+++ src/main/java/com/backend/tienda/repository/ProveedoresRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.Proveedores;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProveedoresRepository extends JpaRepository<Proveedores, Long> {
+public interface ProveedoresRepository extends JpaRepository<Proveedores, Double> {
 }
