+++ src/main/java/com/backend/tienda/servicios/DetalleVentasServicios.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface DetalleVentasServicios {
     public DetalleVentas actualizarDetalleVentas(DetalleVentas detalleVentas);
-    public void eliminarDetalleVentas(Long codigo_detalle_venta);
+    public void eliminarDetalleVentas(Double codigo_detalle_venta);
     public DetalleVentas guardarDetalleVentas(DetalleVentas detalleVentas);
     public List<DetalleVentas> listarDetalleVentas();
-    public DetalleVentas buscarDetalleVenta(Long codigo_detalle_venta);
+    public DetalleVentas buscarDetalleVenta(Double codigo_detalle_venta);
 }
