+++ src/main/java/com/backend/tienda/servicios/ProveedoresServicios.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ProveedoresServicios {
     public Proveedores actualizarProveedores(Proveedores proveedores);
-    public void eliminarProveedores(Long nitproveedor);
+    public void eliminarProveedores(Double nitproveedor);
     public Proveedores guardarProveedores(Proveedores proveedores);
     public List<Proveedores> listarProveedores();
-    public Proveedores buscarProveedor(Long nitproveedor);
+    public Proveedores buscarProveedor(Double nitproveedor);
 }
